Rename ChartDrawer to CSVChartDrawer
