+++ python/test/lib/test_chart_drawer.py
@@ -2,7 +2,7 @@ import datetime
 import json
 import os
 
-from chart_drawer import ChartDrawer
+from csv_chart_drawer import CSVChartDrawer
 from list_handler import __csv_to_dicts__
 
 
@@ -10,12 +10,12 @@ def test_csv(tmp_dir: str) -> None:
     csv_path = os.path.join('..', 'csv', 'test_data.csv')
     with open(os.path.join('..', 'json', 'res_bodies.json')) as f:
         res_bodies = json.loads(f.read())
-    chart_drawer = ChartDrawer(csv_path, res_bodies)
+    csv_chart_drawer = CSVChartDrawer(csv_path, res_bodies)
 
     filename = os.path.join(
         tmp_dir, f'accuracy_score_chart_{datetime.datetime.now():%Y%m%d%H%M%S}.csv',
     )
     with open(filename, 'w') as f:
-        chart_drawer.csv(f)
+        csv_chart_drawer.csv(f)
 
     assert len(__csv_to_dicts__(filename)) == len(__csv_to_dicts__(csv_path))
