+++ ruby-on-rails/spec/lib/chart_drawer_spec.rb
@@ -2,9 +2,9 @@
 
 require 'rails_helper'
 
-RSpec.describe ChartDrawer do
-  let(:chart_drawer) { described_class.new(csv_path, res_bodies) }
-  let(:csv_chart)    { chart_drawer.csv }
+RSpec.describe CSVChartDrawer do
+  let(:csv_chart_drawer) { described_class.new(csv_path, res_bodies) }
+  let(:csv_chart)    { csv_chart_drawer.csv }
   let(:csv_path)     { Rails.root.join('csv', 'test_data.csv') }
   let(:json_path)    { Rails.root.join('json', 'res_bodies.json') }
   let(:test_data)    { CSV.read(csv_path) }
