+++ ruby/.rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2026-05-17 19:31:12 UTC using RuboCop version 1.86.2.
+# on 2026-05-17 19:45:25 UTC using RuboCop version 1.86.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -19,19 +19,13 @@ Layout/EmptyLinesAroundClassBody:
 # SupportedStyles: native, lf, crlf
 Layout/EndOfLine:
   Exclude:
-    - 'src/lib/chart_drawer.rb'
+    - 'src/lib/csv_chart_drawer.rb'
 
 # Offense count: 2
 # This cop supports unsafe autocorrection (--autocorrect-all).
 Lint/NonAtomicFileOperation:
   Exclude:
     - 'test/application_test.rb'
-
-# Offense count: 1
-# This cop supports safe autocorrection (--autocorrect).
-Lint/UselessAssignment:
-  Exclude:
-    - 'test/training_data_formatter_test.rb'
 
 # Offense count: 2
 # Configuration parameters: AllowedMethods, AllowedPatterns, CountRepeatedAttributes.
@@ -64,7 +58,7 @@ Performance/RedundantSplitRegexpArgument:
 # AllowedMethods: lambda, proc, it
 Style/BlockDelimiters:
   Exclude:
-    - 'src/lib/chart_drawer.rb'
+    - 'src/lib/csv_chart_drawer.rb'
     - 'src/lib/format.rb'
     - 'src/queries/accuracy_check_query.rb'
 
@@ -75,7 +69,7 @@ Style/Documentation:
     - 'spec/**/*'
     - 'test/**/*'
     - 'src/accuracy_reporter.rb'
-    - 'src/lib/chart_drawer.rb'
+    - 'src/lib/csv_chart_drawer.rb'
     - 'src/lib/format.rb'
     - 'src/queries/accuracy_check_query.rb'
     - 'src/training_data_formatter.rb'
@@ -84,7 +78,7 @@ Style/Documentation:
 # This cop supports safe autocorrection (--autocorrect).
 Style/EmptyLiteral:
   Exclude:
-    - 'src/lib/chart_drawer.rb'
+    - 'src/lib/csv_chart_drawer.rb'
 
 # Offense count: 1
 # This cop supports safe autocorrection (--autocorrect).
@@ -99,7 +93,7 @@ Style/FetchEnvVar:
 # SupportedStyles: format, sprintf, percent
 Style/FormatString:
   Exclude:
-    - 'src/lib/chart_drawer.rb'
+    - 'src/lib/csv_chart_drawer.rb'
 
 # Offense count: 9
 # This cop supports unsafe autocorrection (--autocorrect-all).
@@ -110,7 +104,7 @@ Style/FrozenStringLiteralComment:
     - '**/*.arb'
     - 'Gemfile'
     - 'src/accuracy_reporter.rb'
-    - 'src/lib/chart_drawer.rb'
+    - 'src/lib/csv_chart_drawer.rb'
     - 'src/lib/format.rb'
     - 'src/queries/accuracy_check_query.rb'
     - 'src/training_data_formatter.rb'
